+++ lib/formatters/masked_input_formatter.dart
@@ -24,17 +24,19 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
+import 'dart:math';
+
 import 'package:flutter/services.dart';
 
-import 'formatter_utils.dart';
-
 class MaskedInputFormatter extends TextInputFormatter {
-  final String? mask;
+  final String mask;
 
   final String _anyCharMask = '#';
   final String _onlyDigitMask = '0';
-  final RegExp? anyCharMatcher;
-  String _lastValue = '';
+  final RegExp? allowedCharMatcher;
+  List<int> _separatorIndices = <int>[];
+  List<String> _separatorChars = <String>[];
+  String _maskedValue = '';
 
   /// [mask] is a string that must contain # (hash) and 0 (zero)
   /// as maskable characters. # means any possible character,
@@ -45,83 +47,139 @@ class MaskedInputFormatter extends TextInputFormatter {
   /// 000-000 will only match digits and won't allow a string like Gtt-RBB
   ///
   /// will match literally any character unless
-  /// you supply an [anyCharMatcher] parameter with a RegExp
+  /// you supply an [allowedCharMatcher] parameter with a RegExp
   /// to constrain its values. e.g. RegExp(r'[a-z]+') will make #
   /// match only lowercase latin characters and everything else will be
   /// ignored
   MaskedInputFormatter(
     this.mask, {
-    this.anyCharMatcher,
-  }) : assert(mask != null);
+    this.allowedCharMatcher,
+  });
 
-  bool get isFilled => mask!.length == _lastValue.length;
+  bool get isFilled => _maskedValue.length == mask.length;
+
+  String get unmaskedValue {
+    _prepareMask();
+    final stringBuffer = StringBuffer();
+    for (var i = 0; i < _maskedValue.length; i++) {
+      var char = _maskedValue[i];
+      if (!_separatorChars.contains(char)) {
+        stringBuffer.write(char);
+      }
+    }
+    return stringBuffer.toString();
+  }
 
   @override
   TextEditingValue formatEditUpdate(
     TextEditingValue oldValue,
     TextEditingValue newValue,
   ) {
-    final bool isErasing = newValue.text.length < oldValue.text.length;
-
-    if (isErasing || _lastValue == newValue.text) {
-      _lastValue = newValue.text;
-      return newValue;
-    }
-
-    final String masked = applyMask(
+    final FormattedValue formattedValue = applyMask(
       newValue.text,
     );
-    final end = newValue.text.length - newValue.selection.end;
+    var selectionOffset = newValue.selection.end + formattedValue.selectionOffset;
+    _maskedValue = formattedValue.toString();
+    if (selectionOffset > _maskedValue.length) {
+      selectionOffset = _maskedValue.length;
+    }
 
-    _lastValue = masked;
     return TextEditingValue(
-      text: masked,
+      text: _maskedValue,
       selection: TextSelection.collapsed(
-        offset: masked.length - end,
+        offset: selectionOffset,
       ),
     );
   }
 
   bool _isMatchingRestrictor(String character) {
-    if (anyCharMatcher == null) {
+    if (allowedCharMatcher == null) {
       return true;
     }
-    return anyCharMatcher!.stringMatch(character) != null;
+    return allowedCharMatcher!.stringMatch(character) != null;
   }
 
-  String applyMask(String text) {
-    final List<String> chars = text.split('');
-    final List<String> result = <String>[];
-
-    int maskShift = 0;
-    for (int i = 0; i < mask!.length; i++) {
-      final int maskIndex = i + maskShift;
-      if (chars.length <= i || mask!.length <= maskIndex) break;
-      final String currentChar = chars[i];
-      final maskChar = mask![maskIndex];
-      if (currentChar == maskChar) {
-        result.add(currentChar);
-      } else if (maskChar == _anyCharMask) {
-        if (_isMatchingRestrictor(currentChar)) {
-          result.add(currentChar);
-        } else {
-          break;
-        }
-      } else if (maskChar == _onlyDigitMask) {
-        if (isDigit(currentChar)) {
-          result.add(currentChar);
-        } else {
-          break;
-        }
-      } else {
-        result.add(maskChar);
-        if (_isMatchingRestrictor(currentChar)) {
-          result.add(currentChar);
-          maskShift++;
+  void _prepareMask() {
+    if (_separatorIndices.isEmpty) {
+      for (var i = 0; i < mask.length; i++) {
+        if (mask[i] != _anyCharMask && mask[i] != _onlyDigitMask) {
+          _separatorIndices.add(i);
+          _separatorChars.add(mask[i]);
         }
       }
     }
+  }
 
-    return result.join();
+  int _countSeparators(String text) {
+    _prepareMask();
+    var numSeparators = 0;
+    for (var i = 0; i < text.length; i++) {
+      if (_separatorChars.contains(text[i])) {
+        numSeparators++;
+      }
+    }
+    return numSeparators;
+  }
+
+  String _removeSeparators(String text) {
+    var stringBuffer = StringBuffer();
+    for (var i = 0; i < text.length; i++) {
+      var char = text[i];
+      if (!_separatorChars.contains(char)) {
+        stringBuffer.write(char);
+      }
+    }
+    return stringBuffer.toString();
+  }
+
+  FormattedValue applyMask(String text) {
+    _prepareMask();
+    int numSeparatorsBefore = _countSeparators(text);
+    String clearedValueAfter = _removeSeparators(text);
+    final isErasing = _maskedValue.length > text.length;
+    FormattedValue formattedValue = FormattedValue();
+    StringBuffer stringBuffer = StringBuffer();
+    var index = 0;
+    var maxLength = min(
+      clearedValueAfter.length,
+      mask.length - _separatorChars.length,
+    );
+
+    for (var i = 0; i < maxLength; i++) {
+      var curChar = clearedValueAfter[i];
+      if (!_isMatchingRestrictor(curChar)) {
+        continue;
+      }
+      if (_separatorIndices.contains(index)) {
+        stringBuffer.write(mask[index]);
+        index++;
+      }
+      stringBuffer.write(curChar);
+      index++;
+    }
+    formattedValue._isErasing = isErasing;
+    formattedValue._formattedValue = stringBuffer.toString();
+    formattedValue._numSeparatorsAfter = _countSeparators(
+      formattedValue._formattedValue,
+    );
+    formattedValue._numSeparatorsBefore = numSeparatorsBefore;
+    return formattedValue;
+  }
+}
+
+class FormattedValue {
+  String _formattedValue = '';
+  int _numSeparatorsBefore = 0;
+  int _numSeparatorsAfter = 0;
+  bool _isErasing = false;
+
+  int get selectionOffset {
+    final offset = (_numSeparatorsAfter - _numSeparatorsBefore).abs();
+    return _isErasing ? 0 : offset;
+  }
+
+  @override
+  String toString() {
+    return _formattedValue;
   }
 }